Rename gemini.ts to api.ts
The Gemini API dependency was removed in PR #1. The file was just a thin
wrapper calling the local /api/extract endpoint with no Gemini code left.
Renamed to src/lib/api.ts and updated the import in useContentData.ts.
